Categorias mas stock y se simplifico el script de la db

+++ GO_WEB/GO_WEB/GO_WEB/Controllers/CursoController.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
+using System.Web.Security;
 
 namespace MN_WEB.Controllers
 {
@@ -23,8 +24,19 @@ namespace MN_WEB.Controllers
         [HttpGet]
         public ActionResult Agregar()
         {
+            var Categoria = model.ConsultarCategoria();
+            var listaCategoria = new List<SelectListItem>();
+
+            foreach (var item in Categoria)
+            {
+                listaCategoria.Add(new SelectListItem { Text = item.NombreCategoria, Value = item.IdCategoria.ToString() });
+            }
+
+            ViewBag.ListaCategoria = listaCategoria;
+
             return View();
         }
+
 
 
         [HttpPost]
@@ -48,26 +60,54 @@ namespace MN_WEB.Controllers
         public ActionResult Editar(long q)
         {
             var datos = model.ConsultarProducto(q);
+            var Categoria = model.ConsultarCategoria();
+            var listaCategoria = new List<SelectListItem>();
+
+            foreach (var item in Categoria)
+            {
+                listaCategoria.Add(new SelectListItem { Text = item.NombreCategoria, Value = item.IdCategoria.ToString() });
+            }
+
+            ViewBag.ListaCategoria = listaCategoria;
+
             return View(datos);
         }
+
 
         [HttpPost]
         public ActionResult EditarProducto(ProductoEnt entidad, HttpPostedFileBase imagenProducto)
         {
-            System.IO.File.Delete(AppDomain.CurrentDomain.BaseDirectory + entidad.Imagen);
+            if (imagenProducto != null && imagenProducto.ContentLength > 0)
+            {
+                if (!string.IsNullOrEmpty(entidad.Imagen))
+                {
+                    System.IO.File.Delete(AppDomain.CurrentDomain.BaseDirectory + entidad.Imagen);
+                }
 
-            entidad.Imagen = string.Empty;
-            model.ActualizarProducto(entidad);
+                entidad.Imagen = string.Empty;  
 
-            string extensionImagen = Path.GetExtension(Path.GetFileName(imagenProducto.FileName));
-            string rutaGuardarImagenes = AppDomain.CurrentDomain.BaseDirectory + "images\\" + entidad.IdProducto + extensionImagen;
-            imagenProducto.SaveAs(rutaGuardarImagenes);
 
-            entidad.Imagen = "\\images\\" + entidad.IdProducto + extensionImagen;
-            model.ActualizarRuta(entidad);
+                model.ActualizarProducto(entidad);
+
+
+                string extensionImagen = Path.GetExtension(Path.GetFileName(imagenProducto.FileName));
+                string rutaGuardarImagenes = AppDomain.CurrentDomain.BaseDirectory + "images\\" + entidad.IdProducto + extensionImagen;
+                imagenProducto.SaveAs(rutaGuardarImagenes);
+
+                entidad.Imagen = "\\images\\" + entidad.IdProducto + extensionImagen;
+                model.ActualizarRuta(entidad);
+            }
+            else
+            {
+                model.ActualizarProducto(entidad);
+            }
 
             return RedirectToAction("ConsultarMantProductos", "Producto");
         }
+
+
+
+
 
     }
 }
